+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.0.1 (2020-??-??)
+
+* Added balance
+
 # 1.0.0 (2020-08-30)
 
 * First downloadable version of the application